requirements.txt
aiofiles==23.2.1
anyio==4.0.0
anylabeling==0.3.3
appdirs==1.4.4
argcomplete==3.1.2
asgiref==3.7.2
attr==0.3.1
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.3
beautifulsoup4==4.12.2
bleach==5.0.1
boto==2.49.0
boto3==1.16.28
botocore==1.19.28
boxing==0.1.4
Brotli==1.1.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dacite==1.7.0
darkdetect==0.8.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
Django==3.2.20
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.4.2
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.1.3
expiringdict==1.2.2
fiftyone==0.22.1
fiftyone-brain==0.13.2
fiftyone-db==0.4.0
fiftyone-desktop==0.29.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
glob2==0.7
google-api-core==2.11.0
google-auth==2.14.1
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphql-core==3.2.3
grpc-google-iam-v1==0.12.4
grpcio==1.59.0
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmlmin==0.1.12
httpcore==0.18.0
httpx==0.25.0
humanfriendly==10.0
humansignal-drf-yasg==1.21.9
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
ijson==3.2.3
imageio==2.31.5
imgviz==1.7.4
inflate64==0.3.1
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
kaleido==0.2.1
kiwisolver==1.4.5
label-studio==1.9.1.post0
label-studio-converter==0.0.57
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
lazy_loader==0.3
lockfile==0.12.2
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib==3.8.0
mongoengine==0.24.2
motor==3.3.1
mpmath==1.3.0
multivolumefile==0.2.3
natsort==8.4.0
networkx==3.0
nltk==3.6.7
numpy==1.24.3
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
ordered-set==4.0.2
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
pprintpp==0.4.0
priority==2.0.0
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
py7zr==0.20.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.13
pymongo==4.5.0
pyparsing==3.1.1
pyppmd==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.1
pyzstd==0.15.9
qimage2ndarray==1.10.0
rarfile==4.1
redis==3.5.3
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.3.7
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
semver==2.13.0
sentry-sdk==1.32.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.31.1
strawberry-graphql==0.138.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchsummary==1.5.1
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ua-parser==0.18.0
ujson==5.8.0
ultralytics==8.0.197
universal-analytics-python3==1.1.1
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
visdom==0.2.4
voxel51-eta==0.12.0
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.4
wrapt==1.15.0
wsproto==1.2.0
xmljson==0.2.0
xmltodict==0.13.0
